requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
jsonfield==3.1.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.0
python-datauri==0.2.9
pytz==2020.1
sqlparse==0.4.1
